Ensure statistics schema is bootstrapped on startup
